Move several functions for preprocessing to base class of ImageSource
